Fixed compiler warning. (#405)
Part of #177

+++ src/act/alexandria/openmm_xml.cpp
@@ -567,7 +567,8 @@ void OpenMMWriter::addXmlNonbonded(xmlNodePtr                       parent,
                 }
                 break;
             default:
-                fprintf(stderr, "Unknown non-bonded force type %d %s\n", fs.potential(),
+                fprintf(stderr, "Unknown non-bonded force type %d %s\n", 
+                        static_cast<int>(fs.potential()),
                         potentialToString(fs.potential()).c_str());
             }
             if (nullptr != fsPtr)
